rename tutorial to exercise
