+++ docs/tutorials/exercise_1/fibonacci.py
