+++ docs/tutorials/exercise_2/holding_onto_memory.py
