+++ docs/tutorials/exercise_3/lru_cache.py
