+++ docs/tutorials/tests/test_exercise_1.py
@@ -1,5 +1,5 @@
 import pytest
-from tutorial_1.fibonacci import generate_fibonacci_hash
+from exercise_1.fibonacci import generate_fibonacci_hash
 
 # Memory tests
